Migrate leads services and repository from typeorm to prisma

+++ server/src/services/LeadsServices/DeleteLeadService.ts
@@ -1,17 +1,24 @@
-import { getCustomRepository } from 'typeorm'
-
 import { LeadsRepository } from '../../repositories/LeadsRepository/PrismaLeadsRepository'
+
+import { Lead } from '../../prisma/models/Lead'
+import { AppError } from '../../errors/AppError'
 
 interface IRequest {
   leadId: string;
 }
 
 export class DeleteLeadService {
-  public async execute({ leadId }: IRequest): Promise<void> {
-    const leadsRepository = getCustomRepository(LeadsRepository)
+  public async execute({ leadId }: IRequest): Promise<Lead> {
+    const leadsRepository = new LeadsRepository()
 
-    await leadsRepository.deleteLead(leadId)
+    const lead = await leadsRepository.findById(leadId)
 
-    return
+    if(!lead) {
+      throw new AppError('Lead não encontrado.')
+    }
+
+    const deleteResult = await leadsRepository.deleteLead(leadId)
+
+    return deleteResult
   }
 }
